[refactor] consolidate vendored code under internal/; fold doc-coverage into module_test
Two structure tidies from review:

1. Move the vendored jsonrepair runtime from lib/json/internal/jsonrepair
   to the top-level internal/jsonrepair. The repo now has a single internal/
   tree, so all vendored third-party code lives in one place — easier to find
   and to license-audit (which matters as this goes open source). The trade-off
   is wider visibility (module-wide instead of lib/json-only), acceptable for a
   leaf utility with one consumer. Updates the import, codecov/.codacy ignore
   paths, and the CLAUDE.md vendoring guidance (vendored code goes under the
   top-level internal/).

2. Fold the standalone doc_coverage_test.go (one test, 94 lines) into
   module_test.go, whose theme is exactly the builtin-module registry and its
   metadata — TestDocCoverage and its libReadmeDir/moduleSurface helpers now
   live beside TestListBuiltinModules / TestModuleCapabilities, one fewer
   root-level test file.

No behavior change. go.sum unchanged (jsonrepair stays zero-dep); make doc-check
still 234/234; gofmt/vet clean; -race -count=2 and the go1.19 floor green.
