+++ internal/jsonrepair/errors.go
